Add files via upload
This lecture provides an in-depth introduction to the for loop in Python, one of the most commonly used control structures for iterating over data. It’s designed for students in the Bachelor of Science in Business Data Analytics program, aiming to build a strong foundation in Python programming with applications for data analytics tasks. By the end of this lecture's session, students will understand the core concepts of the for loop, gain familiarity with loop syntax, and explore various examples related to business data processing.